requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.1.0
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
multidict==6.0.4
protobuf==4.25.1
yarl==1.9.4
